uses build constraints to ignore test/empty.go on compile; renames test pkg 'donotimport' to 'test'; fixes go get with Go 1.3.1
